adding setFunction for Jacobian Type in LCP solution adding setFunction for tolerance in GlobalCriteria function adding setFunction for maximal Damping steps fixing double<->int mixup in SingleContact

+++ trunk/kernel/mbsim/single_contact.cc
@@ -1344,7 +1344,7 @@ namespace MBSim {
 
   }
 
-  void SingleContact::applyh(int t, int j) {
+  void SingleContact::applyh(double t, int j) {
     WF[1] = cpData[0].getFrameOfReference().getOrientation().col(0) * laN(0);
 
     if (fdf)
